requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
SQLAlchemy==2.0.36
tomli==2.0.2
typing_extensions==4.12.2
Werkzeug==3.1.1
